passage des int en double

+++ src/class/Object.cs
@@ -2,20 +2,23 @@ namespace src.Object
 {
     public class Obj
     {
-        private int _position;
-        public int position{
+        private double _position;
+        public double position
+        {
             get => _position;
             set => _position = value;
         }
 
-        private int _weight;
-        public int weightObj{
+        private double _weight;
+        public double weightObj
+        {
             get => _weight;
             set => _weight = value;
         }
 
         // constructor
-        public void constructorObject(int weightObjP, int positionP){ 
+        public void constructorObject(double weightObjP, double positionP)
+        {
             weightObj = weightObjP;
             position = positionP;
         }
